Rework anchor mode.
The JS component now provides a set of `anchorIndices` which represent the current set of potential subviews to use as anchors when the user scrolls. The ScrollView picks an anchor based on scroll position and scroll direction.

The JS component also provides an `anchorMap` of old indices to new indices. This map tells the ScrollView where to find an anchor after the state change. The ScrollView then adjusts the contentOffset so that the previous anchor is at the same position on the screen after the state change.