networking: Initialize NetworkingModule with proper OkHttpClient.
`getOkHttpClient` calls `createClient` if `OkHttpClientProvider` has
no sClient (the root client) yet. Otherwise, it returns the root client.
Thus, it has basically the same effect as `createClient` in this case,
since only one NetworkingModule is ever instantiated and sharing its
client with the very few other instances where OkHttpClientProvider is
used shouldn't be a problem.

We need this change so that OkHttpClientProvider gets an initialized
sClient (`createClient` doesn't initialize the sClient, it just returns
a new client). This will be necessary in the next commit, which fixes
a networking bug.